requirements.txt
tensorflow-gpu
tensor2tensor
requests
clint